#77 refactor: 디렉터리 구조 리펙토링
