+++ frontend/src/pages/About/Faculty/FacultyStyle.ts
