+++ frontend/src/pages/About/Faculty/ProfessorCard.tsx
